Add files via upload
Added :
PopupCreator.java - it's only purpose is to create new objects of the RecipePopup class.

Modified:
MainMenuScreen.createScrollPane() -  to get the popup making process  started, a popupManager object is created here. ObjectPool and popup creation is elsewhere now
PopupManager -  no longer takes in an ObjectPool, it gets the poolInstance.; it also communes with the Object Pool to begin "getting" popups to display
ObjectPool .removeObject() renamed  to reuseObject() and modified to work in concert with waitForObject() which is also modified to work better with the PopupManager.
RecipePopup.returnIt()  - added to facilitate object release to the pool
RecipePopup.popupImage() - dissected to breakout portions of the code that are important to object reuse.  
RecipePopup..getURL() - provides the image retrieval information
RecipePopup..setPic() -  sets the pictures for each drink popup; and because it's returned variable is global now, it can be used for reuse
RecipePopup.refreshPane() -  used strictly for reused popup objects - refreshes the information in the target popup with the data of the passed in recipe
RecipePopup class variables textAreaI, textAread, rName, popupPic are declared globally now to allow visibility to refreshPane() .

ToDo (maybe):
Migrate the favoriting functionality to .refreshPane()